clantastic/python/requirements.txt
unicornhat
paho-mqtt==1.2
python-dotenv==0.6.3
